smart-support-hub/requirements.txt

streamlit==1.38.0
pandas==2.2.2
pydantic==2.7.4
python-dotenv==1.0.1
google-generativeai==0.7.2
google-auth==2.31.0
gspread==6.1.2
authlib==1.3.2
requests==2.32.3
tenacity==8.3.0
ruff==0.5.7
black==24.8.0
pytest==8.3.2
